feat: move install.sh to scripts folder

+++ scripts/local.sh
@@ -26,8 +26,12 @@ vendor/bin/sail build
 echo -e "\n${bold}> sail up -d${normal}"
 vendor/bin/sail up -d
 
-echo -e "\n${bold}> sail artisan octane:install --server=frankenphp${normal}"
-vendor/bin/sail artisan octane:install --server=frankenphp --force
+if [ ! -f ./frankenphp ]; then
+    echo -e "\n${bold}> sail artisan octane:install --server=frankenphp${normal}"
+    [ -f ./config/octane.php ] && cp ./config/octane.php ./config/octane.php.bak
+    vendor/bin/sail artisan octane:install --force --server=frankenphp --force
+    [ -f ./config/octane.php.bak ] && mv ./config/octane.php.bak ./config/octane.php
+fi
 
 echo -e "\n${bold}> sail artisan storage:link${normal}"
 vendor/bin/sail artisan storage:link
